Abstraction of base outputs, added ase and pymatgen files where the conversions should take place. In all other files, ase and pymatgen should not be used.
