--- src/qe_tools/converters/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from qe_tools.converters.structure import get_parameters_from_cell
-
-__all__ = ('get_parameters_from_cell',)
